Always emit initialization function
On some occasions we may not need initialization function but already
have some LLVM instructions been created for purpose of module
initialization. In such case we must get rid of them because (a) these
instructions are not contained in any basic block and (b) these
instructions may reference some global value in the module. On such
occastions module will not pass verification and we'll crash.

The easiest fix seems to be internalizing of initialization function in
case we don't really need it, so LLVM optimization passes will eradicate
it and verifier is happy.