Don't inline function if previous declaration is different
It is possible that some of modules contains newer functionality then
our standard library does and compiler can do different inlining
decisions. For example the same function may not be inlined in standard
library but can be inlined in newer version of some module from standard
library which we import with -fgo-importcfg. In such case we do avoid
inlining, because we can face an error due to missing result variables
in original declaration.