Analyze size parameter in calls to runtime.mallocgc
The runtime.mallocgc function may be used to allocate arrays of
elements, in this case size may be different from object size obtained
from type descriptor. When size is not constant we should not try to
optimize such allocation, because we don't know amount of memory to
allocate on stack.