Replace calls to runtime.gcWriteBarrier with store instruction
On some occasions we can't just remove a call to runtime.gcWriteBarrier
when optimizing out GC heap allocation. This happens because gofrontend
is not always generating a follow-up store instruction. Instead of
removing a call we now generate a store instruction.

TODO: We need to investigate what passes should be run after
gogcoptimize to eliminate redundand basic blocks